Build Streamlit frontend for AI Research Assistant

+++ app.py
@@ -2,7 +2,7 @@ from fastapi import FastAPI, UploadFile, File, HTTPException, Depends
 from RAG import get_response
 from pydantic import BaseModel, Field
 import sqlite3
-from database import load_chat_history, save_chat_history, load_and_chunk_pdf_file, save_pdf_file, update_vectorstore, get_chat_history, get_documents, delete_document, document_exists, create_user, username_exists, get_user_by_username, get_connection
+from database import load_chat_history, save_chat_history, load_and_chunk_pdf_file, save_pdf_file, update_vectorstore, get_chat_history, get_documents, delete_document, document_exists, create_user, username_exists, get_user_by_username, get_connection, create_projects, get_projects, delete_project, verify_project_owner
 from auth import hash_password, verify_password, create_access_token, get_current_user
 from fastapi.security import OAuth2PasswordRequestForm
 from typing import Annotated
@@ -60,30 +60,68 @@ def login_user(form_data: OAuth2PasswordRequestForm = Depends()):
             "token_type": "bearer"
       }
        
-       
+@app.post('/projects', description='Create a new project here')
+def register_project(current_user: Annotated[dict, Depends(get_current_user)], project_name: str):
+
+         user_id = current_user['user_id']   
+         project_id = create_projects(user_id, project_name)
+         return {
+                'message': 'Project successfully registered', 
+                'Project ID': project_id,
+                'Project Name': project_name
+                }
+
+@app.get('/current_projects', description='View all the projects details you own.')
+def get_current_projects(current_user: Annotated[dict, Depends(get_current_user)]):
+
+       user_id = current_user['user_id']
+       return get_projects(user_id)
+
+@app.delete('/deleteproject', description='chokeslam the shit out of a project')
+def kill_project(current_user: Annotated[dict, Depends(get_current_user)], project_id: int):
+
+       user_id = current_user['user_id']
+       if verify_project_owner(project_id, user_id):
+              delete_project(project_id, user_id)
+              return {
+                     'message': 'Project deleted successfully'
+              }
+
+       else:
+              raise HTTPException(
+                     status_code=403, 
+                     detail='Unauthorized access!'
+              )
+
+
 
 @app.post('/upload', description='Upload your PDF File here')
-async def upload_and_save_pdf(file: UploadFile = File(...), current_user:dict = Depends(get_current_user)):
+async def upload_and_save_pdf(project_id: int, file: UploadFile = File(...), current_user:dict = Depends(get_current_user)):
 
         if file.content_type != 'application/pdf':
             raise HTTPException(status_code=400, detail='Please upload a valid PDF file')
         
         user_id = current_user['user_id']
-
+        if not verify_project_owner(project_id, user_id):
+               raise HTTPException(
+                      status_code=403,
+                      detail='Unauthorized access'
+               )
+        
         # save PDF file into documents folder and documents table:
-        if document_exists(user_id, file.filename):
+        if document_exists(project_id, file.filename):
                     raise HTTPException(
                           status_code=409,
                           detail="This document has already been uploaded."
                     )
         
         
-        document_id, filepath, filename = await save_pdf_file(file, user_id)
+        document_id, filepath, filename = await save_pdf_file(file, project_id)
 
             
         
         # load chunks in chunks table:
-        chunks = load_and_chunk_pdf_file(user_id, document_id, filepath, filename)
+        chunks = load_and_chunk_pdf_file(project_id, document_id, filepath, filename)
 
         # create embeddings and load in vectorstore:
         update_vectorstore(chunks)
@@ -91,36 +129,57 @@ async def upload_and_save_pdf(file: UploadFile = File(...), current_user:dict = 
         return {'message': 'PDF saved successfully', 'filename': filename, 'filepath': filepath, 'document ID': document_id, 'Number of Chunks': len(chunks)}
 
 @app.post('/chat', description='Start asking questions about uploaded PDFs and save responses in a chatbox')
-def user_llm_convo(question: str, current_user: dict = Depends(get_current_user)):
+def user_llm_convo(project_id: int, question: str, current_user: dict = Depends(get_current_user)):
 
       user_id = current_user['user_id']
-      response = get_response(user_id, question)
+      if not verify_project_owner(project_id, user_id):
+             raise HTTPException(
+                    status_code=403, 
+                    detail='Unauthorized access'
+             )
+      response = get_response(project_id, question)
       return response
 
 @app.get('/chathistory', description='Get the entire chat thread you have communicated with the LLM till now')
-def return_chat_history(current_user: dict = Depends(get_current_user)):
+def return_chat_history(project_id: int, current_user: dict = Depends(get_current_user)):
       user_id = current_user['user_id']
-      return {'history': get_chat_history(user_id)}
+      if not verify_project_owner(project_id, user_id):
+             raise HTTPException(
+                    status_code=403,
+                    detail='Unauthorized access'
+             )
+      return {'history': get_chat_history(project_id)}
 
 @app.delete('/deletechat', description='delete the chat history and start a fresh conversation')
-def delete_chat(current_user: Annotated[dict, Depends(get_current_user)]):
+def delete_chat(project_id: int, current_user: Annotated[dict, Depends(get_current_user)]):
 
       user_id = current_user['user_id']
+      if not verify_project_owner(project_id, user_id):
+             raise HTTPException(
+                    status_code=403,
+                    detail='Unauthorized access'
+             )
       conn = get_connection()
       cursor = conn.cursor()
       SQL_query = """
-            DELETE FROM chat_history WHERE user_id = ?
+            DELETE FROM chat_history WHERE project_id = ?
         """
-      cursor.execute(SQL_query, (user_id,))
+      cursor.execute(SQL_query, (project_id,))
       conn.commit()
       conn.close()
 
       return {'message':'Chat history deleted succesfully'}
 
 @app.get('/documents', description='fetches the PDF file name and associated document ID')
-def get_docs(current_user: dict = Depends(get_current_user)):
+def get_docs(project_id: int, current_user: dict = Depends(get_current_user)):
+
       user_id = current_user['user_id']
-      rows = get_documents(user_id)
+      if not verify_project_owner(project_id, user_id):
+             raise HTTPException(
+                    status_code=403,
+                    detail='Unauthorized access'
+             )
+      rows = get_documents(project_id)
       result = []
       for row in rows:
             result.append({
@@ -131,9 +190,14 @@ def get_docs(current_user: dict = Depends(get_current_user)):
       return result
 
 @app.delete('/deletedocuments', description='Delete any document by referencing the document ID')
-def delete_docss(current_user: Annotated[dict, Depends(get_current_user)], document_id: int):
+def delete_docss(project_id: int, current_user: Annotated[dict, Depends(get_current_user)], document_id: int):
 
       user_id = current_user['user_id']
-      delete_document(user_id, document_id)
+      if not verify_project_owner(project_id, user_id):
+             raise HTTPException(
+                    status_code=403,
+                    detail='Unauthorized access'
+             )
+      delete_document(project_id, document_id)
       return {'message': 'Document deleted succesfully!'}
